Move all non-main service files into a sub-module
