+++ dlite_entities_service/service/uvicorn.py
